Add instruction for applying legacy styles
Anyone upgrading this lib to v1 will lose all their list styles.  It's a significant headache to figure out there are no styles anymore and then recreate them.

This comment or a similar one will help a lot of people I expect, whether they are upgrading or just new to the library and want some sane list styles out of the box.

Thanks!