fix(SearchBar|Input): Allow height on fields to scale with fontSize
Inputs were set a fixed height. Meaning if the developer increase
the fontSize of the input, it would result in a scrollable area
in the text field. Also if users had on display scaling for
accessibility reasons, then they'd be unable to view the text.

This PR changes the fixed height to minHeight to allow scaling.

Fixes #1239