feat(Image): Allow children to be displayed while image loads
Since
https://github.com/react-native-training/react-native-elements/commit/e2795821fa85bbce48282b9c92d2260658d5c35e
the Image component supports nesting children being based from
ImageBackground. This meant however than children of the Image would not
be visible until the image loaded which is incorrect behaviour.

This commit fixes this, allowing children to be visible while the image loads
in the background. Also instead of using the ImageBackground component
from react native, we implement our own based on the source. This is so
we can place the animated placeholder between the content and the image.

e.g (background -> animated placeholder -> our content)

Also makes some small tweaks to Tile and FeaturedTile so that they
display correctly with the new changes to Image.

Since Tile, Avatar, and Card now use RNE Image, the docs and types have
been updated to reflect.