Add `underlayColor` to `HeaderIcon` type
This change will let users specify an `underlayColor` while using an icon in their `Header`. As an example, when faced with a problem like in https://github.com/react-native-elements/react-native-elements/issues/640 we can either use an `Icon` component and pass it the `underlayColor` prop or add the `underlayColor` to the `rightComponent` `HeaderIcon` type. However, the latter is not currently supported by the type definitions.