Fix customisation page documentation
Copy pasting the explanation to extend the theme does not work as expected: it overrides the initial declaration instead of extending it. (eg. "EveryComponent" does not exist in "react-native-elements")
To fix that you need to first import the package then declare again. Adding the import statement on the top does achieve the expected behavior.